requirements.txt
requests
influxdb
influxdb-client
paho-mqtt